+++ CHANGELOG.md
@@ -55,6 +55,12 @@
 * Fix missing target features in module when enabling reference types or multi-value transformation.
   [#3967](https://github.com/rustwasm/wasm-bindgen/pull/3967)
 
+* Fixed Rust values getting GC'd while still borrowed.
+  [#3940](https://github.com/rustwasm/wasm-bindgen/pull/3940)
+
+* Fixed Rust values not getting GC'd if they were created via. a constructor.
+  [#3940](https://github.com/rustwasm/wasm-bindgen/pull/3940)
+
 --------------------------------------------------------------------------------
 
 ## [0.2.92](https://github.com/rustwasm/wasm-bindgen/compare/0.2.91...0.2.92)